LayoutTests/ietestcenter/Javascript/15.4.4.15-5-1-expected.txt

15.4.4.15-5-1 (TestCases/chapter15/15.4/15.4.4/15.4.4.15/15.4.4.15-5-1.js) Array.prototype.lastIndexOf when fromIndex is string

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ES5Harness.preconditionPassed is true
FAIL ES5Harness.testPassed should be true (of type boolean). Was undefined (of type undefined).
PASS successfullyParsed is true

TEST COMPLETE
